pyproject.toml
[tool.poetry]
name = "ai-workflow"
version = "0.1.0"
description = ""
authors = ["Ayman Mahmoud"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
gradio = "^5.15.0"
pypdf = "^5.2.0"
pandas = "^2.2.3"
semantic-kernel = "^1.23.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
